requirements.txt
pyautogui
pillow
transformers
torch
discord.py
pynput